+++ oauth2_provider/settings.py
@@ -41,6 +41,7 @@ ACCESS_TOKEN_MODEL = getattr(settings, "OAUTH2_PROVIDER_ACCESS_TOKEN_MODEL", "oa
 ID_TOKEN_MODEL = getattr(settings, "OAUTH2_PROVIDER_ID_TOKEN_MODEL", "oauth2_provider.IDToken")
 GRANT_MODEL = getattr(settings, "OAUTH2_PROVIDER_GRANT_MODEL", "oauth2_provider.Grant")
 REFRESH_TOKEN_MODEL = getattr(settings, "OAUTH2_PROVIDER_REFRESH_TOKEN_MODEL", "oauth2_provider.RefreshToken")
+SESSION_MODEL = getattr(settings, "OAUTH2_PROVIDER_SESSION_MODEL", "oauth2_provider.Session")
 PAR_REQUEST_MODEL = getattr(
     settings, "OAUTH2_PROVIDER_PAR_REQUEST_MODEL", "oauth2_provider.PushedAuthorizationRequest"
 )
@@ -91,10 +92,12 @@ DEFAULTS = {
     "DEVICE_GRANT_MODEL": DEVICE_GRANT_MODEL,
     "GRANT_MODEL": GRANT_MODEL,
     "REFRESH_TOKEN_MODEL": REFRESH_TOKEN_MODEL,
+    "SESSION_MODEL": SESSION_MODEL,
     "PAR_REQUEST_MODEL": PAR_REQUEST_MODEL,
     # Admin classes
     "APPLICATION_ADMIN_CLASS": "oauth2_provider.authorization_server.admin.ApplicationAdmin",
     "AUTHORIZATION_ADMIN_CLASS": "oauth2_provider.authorization_server.admin.AuthorizationAdmin",
+    "SESSION_ADMIN_CLASS": "oauth2_provider.authorization_server.admin.SessionAdmin",
     "ACCESS_TOKEN_ADMIN_CLASS": "oauth2_provider.authorization_server.admin.AccessTokenAdmin",
     "GRANT_ADMIN_CLASS": "oauth2_provider.authorization_server.admin.GrantAdmin",
     "ID_TOKEN_ADMIN_CLASS": "oauth2_provider.authorization_server.admin.IDTokenAdmin",
@@ -339,6 +342,7 @@ IMPORT_STRINGS = (
     "SCOPES_BACKEND_CLASS",
     "APPLICATION_ADMIN_CLASS",
     "AUTHORIZATION_ADMIN_CLASS",
+    "SESSION_ADMIN_CLASS",
     "ACCESS_TOKEN_ADMIN_CLASS",
     "GRANT_ADMIN_CLASS",
     "ID_TOKEN_ADMIN_CLASS",